requirements.txt
PySide6
FastAPI
uvicorn
SQLAlchemy
requests
python-dotenv
pydantic
